Add Copy text option to wave drop mobile menu
On touch devices text selection in waves/DMs is intentionally disabled
to support the long-press action sheet, but the sheet only offered
"Copy link" - there was no way to copy a message's text on mobile.

Adds a "Copy text" item to the action sheet (guest and authenticated
menus). The message formatting logic is extracted from
useWaveDropsClipboard into helpers/waves/drop-clipboard.helpers.ts so
mobile copy produces the same output as desktop Ctrl+C copy (author +
timestamp heading, reply/quote context, embeds, attachment URLs).
